Moved Tests around.
