+++ src/test/java/net/tullco/addressbook/test_suites/UtilTests.java
@@ -7,9 +7,10 @@ import org.junit.runners.Suite.SuiteClasses;
 import net.tullco.addressbook.test_classes.LocaleUtilTest;
 import net.tullco.addressbook.test_classes.SQLUtilTest;
 import net.tullco.addressbook.test_classes.SystemUtilTest;
+import net.tullco.addressbook.test_classes.ViewUtilTest;
 
 @RunWith(Suite.class)
-@SuiteClasses({ SQLUtilTest.class, SystemUtilTest.class, LocaleUtilTest.class })
+@SuiteClasses({ SQLUtilTest.class, SystemUtilTest.class, LocaleUtilTest.class, ViewUtilTest.class })
 public class UtilTests {
 
 }
